refactor: rename file and use Enum
