+++ api/api.ts
@@ -1,9 +1,11 @@
 import axios from "axios";
 
-export const chatRoom = "chat-room";
-export const profile = "profile";
-export const login = "login";
-export const signup = "signup";
+export enum pathList {
+  ChatRoom = "chat-room",
+  Profile = "profile",
+  Login = "login",
+  Signup = "signup",
+}
 
 export const client = axios.create({
   baseURL: `${process.env.NEXT_PUBLIC_API_URL}/`,
