+++ api/chatrooms/index.ts
@@ -1,11 +1,13 @@
-import { chatRoom, client } from "@api/endpoints";
+import { client, pathList } from "@api/api";
 import { ChatroomsListType, NewChatroomType } from "@type/ChatroomsList";
 import axios from "axios";
 
 export const fetchChatroomsList = async (): Promise<
   Array<ChatroomsListType>
 > => {
-  const response = await client.get<Array<ChatroomsListType>>(`${chatRoom}`);
+  const response = await client.get<Array<ChatroomsListType>>(
+    pathList.ChatRoom
+  );
 
   if (response.status !== 200) {
     throw new Error("failed to submit login info");
@@ -16,10 +18,7 @@ export const fetchChatroomsList = async (): Promise<
 export const createChatroom = async (
   title: string
 ): Promise<NewChatroomType> => {
-  const response = await axios.post<NewChatroomType>(
-    `${process.env.NEXT_PUBLIC_API_URL}/chat-rooms`,
-    title
-  );
+  const response = await axios.post<NewChatroomType>(pathList.ChatRoom, title);
 
   if (response.status !== 200) {
     alert("채팅방 생성에 실패했습니다");
